keeping CDash happy
git-svn-id: https://svn.opendtect.org/od/trunk@34217 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiWell/uiwelllogcalc.cc
@@ -436,7 +436,7 @@ bool uiWellLogCalc::getInpDatas( Well::LogSet& wls,
 	if ( form_.isConst(iinp) )
 	{
 	    InpData inpd; inpd.isconst_ = true;
-	    inpd.constval_ = form_.getConstVal( iinp );
+	    inpd.constval_ = (float)form_.getConstVal( iinp );
 	    if ( mIsUdf(inpd.constval_) )
 		mErrRet(BufferString("Please enter a value for ",
 					form_.variableName(iinp)))
